+++ src/briefy/leica/models/professional/__init__.py
@@ -4,7 +4,6 @@ from briefy.common.db.mixins.optin import OptIn
 from briefy.leica.db import Base
 from briefy.leica.models import mixins
 from briefy.leica.models.professional.workflows import ProfessionalWorkflow
-from briefy.leica.models.professional.workflows import SkillWorkflow
 from briefy.ws.utils.user import add_user_info_to_state_history
 from sqlalchemy import orm
 from sqlalchemy.ext.declarative import declared_attr
@@ -210,7 +209,7 @@ class Professional(ProfessionalMixin, Base):
 class Skill(mixins.LeicaMixin):
     """A Skill of a professional."""
 
-    _workflow = SkillWorkflow
+    _workflow = ProfessionalWorkflow
 
     @declared_attr
     def id(cls):