+++ src/briefy/leica/models/professional/workflows.py
@@ -24,15 +24,84 @@ class ProfessionalWorkflow(BriefyWorkflow):
         'Inserted into the database.'
     )
 
+    pending = WS(
+        'pending', 'Pending',
+        'Waiting for QA Approval.'
+    )
+
+    validation = WS(
+        'validation', 'Legal checking',
+        'Waiting for legal/documents checking.'
+    )
+
+    rejected = WS(
+        'rejected', 'Rejected',
+        'Professional that does not meet the requirements.'
+    )
+
+    trial = WS(
+        'trial', 'On Trial',
+        'Professional that is on trial basis.'
+    )
+
+    active = WS(
+        'active', 'Active',
+        'Professional that is currently working with us.'
+    )
+
+    inactive = WS(
+        'inactive', 'Inactive',
+        'Professional that is not available to be assigned.'
+    )
+
     deleted = WS(
-        'deleted', 'Deleted link',
+        'deleted', 'Deleted Skill',
         'Professional was deleted from the platform.'
     )
 
     # Transitions:
     @created.transition(deleted, 'can_delete')
+    @active.transition(deleted, 'can_delete')
+    @inactive.transition(deleted, 'can_delete')
+    @trial.transition(deleted, 'can_delete')
+    @rejected.transition(deleted, 'can_delete')
+    @validation.transition(deleted, 'can_delete')
+    @pending.transition(deleted, 'can_delete')
     def delete(self):
         """Delete a professional."""
+        pass
+
+    @created.transition(pending, 'can_submit')
+    def submit(self):
+        """Submit a Professional for QA Validation."""
+        pass
+
+    @pending.transition(rejected, 'can_quality_reject')
+    @validation.transition(rejected, 'can_legal_reject')
+    def reject(self):
+        """Reject a professional application."""
+        pass
+
+    @pending.transition(validation, 'can_approve')
+    def approve(self):
+        """Quality approval of a professional application."""
+        pass
+
+    @validation.transition(trial, 'can_validate')
+    def validate(self):
+        """Legal validation of a professional application."""
+        pass
+
+    @trial.transition(active, 'can_activate')
+    @inactive.transition(active, 'can_activate')
+    def activate(self):
+        """Activate a professional."""
+        pass
+
+    @active.transition(inactive, 'can_inactivate')
+    @trial.transition(inactive, 'can_inactivate')
+    def inactivate(self):
+        """Inactivate a professional."""
         pass
 
     @Permission(groups=[LR['owner'], G['scout'], ])
@@ -40,33 +109,39 @@ class ProfessionalWorkflow(BriefyWorkflow):
         """Validate if user can delete this professional."""
         return True
 
-
-class SkillWorkflow(BriefyWorkflow):
-    """Workflow for a Skill."""
-
-    entity = 'skill'
-    initial_state = 'created'
-
-    # States
-    created = WS(
-        'created', 'Created',
-        'Inserted into the database.'
-    )
-
-    deleted = WS(
-        'deleted', 'Deleted Skill',
-        'Professional Skill was deleted from the platform.'
-    )
-
-    # Transitions:
-    @created.transition(deleted, 'can_delete')
-    def delete(self):
-        """Delete a professional."""
-        pass
-
     @Permission(groups=[LR['owner'], G['scout'], ])
-    def can_delete(self):
-        """Validate if user can delete this professional."""
+    def can_submit(self):
+        """Validate if user can submit this professional for QA approval."""
+        return True
+
+    @Permission(groups=[LR['owner'], G['qa'], ])
+    def can_quality_reject(self):
+        """Validate if user can reject this professional application."""
+        return True
+
+    @Permission(groups=[LR['owner'], G['finance'], ])
+    def can_legal_reject(self):
+        """Validate if user can reject this professional application."""
+        return True
+
+    @Permission(groups=[LR['owner'], G['qa'], ])
+    def can_approve(self):
+        """Validate if user can approve this professional application."""
+        return True
+
+    @Permission(groups=[LR['owner'], G['finance'], ])
+    def can_validate(self):
+        """Validate if user can validate this professional application."""
+        return True
+
+    @Permission(groups=[LR['owner'], G['system'], G['pm'], G['qa'], G['scout']])
+    def can_activate(self):
+        """Validate if user can activate this professional application."""
+        return True
+
+    @Permission(groups=[LR['owner'], G['system'], G['pm'], G['qa'], G['scout']])
+    def can_inactivate(self):
+        """Validate if user can inactivate this professional application."""
         return True